// src/elements/public/Admin/types.ts
interface Link {
  href: string;
  title: string;
  type?: string;
}

type Curies = [
  {
    name: string;
    href: string;
    templated: boolean;
  }
];

export interface FxBookmark {
  _links: {
    curies: Curies;
    self: Link;
    'fx:property_helpers': Link;
    'fx:reporting': Link;
    'fx:encode': Link;
    'fx:user': Link;
    'fx:store': Link;
    'fx:stores': Link;
    'fx:token': Link;
  };

  message: string;
}